address comments and refactor code
