+++ backend/services/pg_react_agent_service.py
@@ -14,9 +14,8 @@ from google.genai import types
 from langgraph.graph import StateGraph, END
 
 from services.gemini_service import extract_python_code, thinking_config_for
-from services.pg_react_sql_heuristics import (
+from services.pg_sql_utils import (
     _enrich_schema_context_from_db,
-    _heuristic_sql,
     _normalize_categorical_literals,
     _schema_context_has_columns,
 )
@@ -53,6 +52,9 @@ Rules:
 2. Prefer a SELECT. Include a LIMIT (<= 50) unless the user asks for all rows.
 3. Use only schema-listed tables/columns; never invent names.
 4. Standard PostgreSQL syntax. Use schema-qualified names when given.
+5. When joining tables, follow FK relationships listed in the schema; alias tables a, b, c… in join order.
+6. Interpret "with a <column>" as column IS NOT NULL; "without a <column>" as column IS NULL.
+7. Normalize string literal values to lowercase (e.g. status = 'active', not 'Active').
 """
 
 EVALUATE_PROMPT = """You are a database QA reviewer for a generated PostgreSQL query.
@@ -93,15 +95,6 @@ def _generate(state: _State):
         schema_context = _enrich_schema_context_from_db(
             state.get("conn"), schema_context
         )
-
-    heuristic_sql = _heuristic_sql(state["user_input"], schema_context)
-    if heuristic_sql:
-        heuristic_sql = _normalize_categorical_literals(heuristic_sql, schema_context)
-        return {
-            "generated_query": heuristic_sql,
-            "is_write_action": is_write_sql(heuristic_sql),
-            "iterations": state.get("iterations", 0) + 1,
-        }
 
     prompt = GENERATE_PROMPT.format(
         user_input=state["user_input"],
